// ui/cc-elections-2025/src/types/apiData.ts
export type IndividualCandidate = {
  candidate: {
    id: number;
    candidateType: 'individual',
    name: string;
    email: string;
    country: string;
    socialX: string;
    socialLinkedin: string;
    socialDiscord: string;
    socialTelegram: string;
    socialOther: string;
    publicContact: string;
    about: string;
    bio: string;
    additionalInfo: string;
    videoPresentationLink: string;
    reasonToServe: string;
    governanceExperience: string;
    communicationStrategy: string;
    ecosystemContributions: string;
    legalExpertise: string;
    weeklyCommitmentHours: number;
    conflictOfInterest: string;
    drepId: string;
    stakeId: string;
    createdAt: string;
    updatedAt: string;
    xverification: string;
  }
}